backend/requirements.txt
# StreamFlow Backend Dependencies

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.23

# Video Extraction (auto-updated)
yt-dlp>=2023.12.30

# Scraping
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0

# Browser Automation (auto-updated)
playwright>=1.40.0

# Caching (optional)
redis>=5.0.0

# HTTP Client
httpx>=0.25.0

# Multipart uploads
python-multipart>=0.0.6

# Image Processing
Pillow>=10.0.0

# ====================
# Development / Testing
# ====================

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Linting
ruff>=0.1.0
mypy>=1.7.0

